requirements.txt
# Feedly API client from GitHub
git+https://github.com/feedly/python-api-client.git

# Development dependencies
pytest>=7.0.0
pre-commit>=3.5.0  # For git hooks
black>=23.3.0      # Code formatting
isort>=5.12.0      # Import sorting
ruff>=0.0.275      # Fast Python linter
bandit>=1.7.5      # Security checks

# Additional dependencies
requests>=2.31.0
pymongo>=4.6.0     # MongoDB client
python-dotenv>=1.0.0  # Environment management
pyyaml>=6.0.1      # YAML processing
schedule>=1.2.0    # Task scheduling
openai>=1.86.0     # OpenAI API client
ollama>=0.1.6      # Local LLM client
textblob>=0.17.1   # Text analysis
nltk>=3.8.1        # Natural language processing
numpy>=1.26.0      # Required by NLTK
